Refactor post analysis utilities
