+++ glacium/post/analysis/__init__.py
@@ -1,0 +1,15 @@
+from .cp import read_tec_ascii, compute_cp, plot_cp
+from .ice_thickness import read_wall_zone, process_wall_zone, plot_ice_thickness
+from .ice_contours import load_contours, plot_overlay, animate_growth
+
+__all__ = [
+    "read_tec_ascii",
+    "compute_cp",
+    "plot_cp",
+    "read_wall_zone",
+    "process_wall_zone",
+    "plot_ice_thickness",
+    "load_contours",
+    "plot_overlay",
+    "animate_growth",
+]
